fixed: split non-template class in header/implementation
